First version of additive 2D resampler
Finally some demo of additive 2D resampler integrated into real Avisynth plugin. It supports planar input colorspaces with equal or different sizes of planes (i.e. 4:4:4 to 4:2:0). It currently have only C implemented of additive processing loop and also looks MSVC2019 do not unroll loop or attempt to SIMD it. May be Intel C compiler will. Or it must be manually rewrite to ASM functions.  Unfortunately OpenMP for main processing loop do not starts well with x64 version of binary though works good in x86 2D_resampler separtate project. The pointers are local for rows but if enabling >1 thread it looks pointers are damaged for unknown reason. May be it syntax must be fixed for x64 OpenMP compiling or try Intel C compiler. For speed it looks slower in compare with current JincResize processing at low sizes of frames and up to 3 times faster at large frame sizes - like upsize 3000x2000 2x with taps=10. It currently uses additive processing only on upsizing on equal integer value (like 2,3,4,etc). It have workaround for edges but still not for corners of buffer - so corners may suffer of ringing. May be in futire versions corner of El buffer must be filled with duplicated corner sample if input buffer or may be mirrored diagonally of corner part of input buffer.